+++ config.ru
@@ -1,3 +1,6 @@
 require_relative 'app'
+ROUTES = { '/time' => App.new }.freeze
 
-run App.new
+use Rack::Reloader
+use Rack::ContentType, 'text/plain'
+run Rack::URLMap.new(ROUTES)
